+++ Modules/Microsoft365DSC/DSCResources/MSFT_SPOTenantSettings/MSFT_SPOTenantSettings.psm1
@@ -50,10 +50,6 @@ function Get-TargetResource
 
         [Parameter()]
         [System.Boolean]
-        $UserVoiceForFeedbackEnabled,
-
-        [Parameter()]
-        [System.Boolean]
         $PublicCdnEnabled,
 
         [Parameter()]
@@ -337,10 +333,6 @@ function Set-TargetResource
 
         [Parameter()]
         [System.Boolean]
-        $UserVoiceForFeedbackEnabled,
-
-        [Parameter()]
-        [System.Boolean]
         $PublicCdnEnabled,
 
         [Parameter()]
@@ -502,13 +494,7 @@ function Set-TargetResource
     $CurrentParameters.Remove('AllowSelectSecurityGroupsInSPSitesList') | Out-Null
     $CurrentParameters.Remove('EnableAzureADB2BIntegration') | Out-Null
     $CurrentParameters.Remove('OneDriveSharingCapability') | Out-Null
-
     $CurrentParameters.Remove('TenantDefaultTimezone') | Out-Null # this one is updated separately using Graph
-    if ($CurrentParameters.Keys.Contains('UserVoiceForFeedbackEnabled'))
-    {
-        Write-Verbose -Message 'Property UserVoiceForFeedbackEnabled is deprecated, removing it'
-        $CurrentParameters.Remove('UserVoiceForFeedbackEnabled') | Out-Null
-    }
 
     if ($PublicCdnEnabled -eq $false)
     {
@@ -638,10 +624,6 @@ function Test-TargetResource
         [Parameter()]
         [System.Boolean]
         $UsePersistentCookiesForExplorerView,
-
-        [Parameter()]
-        [System.Boolean]
-        $UserVoiceForFeedbackEnabled,
 
         [Parameter()]
         [System.Boolean]
@@ -793,7 +775,6 @@ function Test-TargetResource
             'LegacyAuthProtocolsEnabled', `
             'SignInAccelerationDomain', `
             'UsePersistentCookiesForExplorerView', `
-            'UserVoiceForFeedbackEnabled', `
             'PublicCdnEnabled', `
             'PublicCdnAllowedFileTypes', `
             'UseFindPeopleInPeoplePicker', `