+++ Modules/Microsoft365DSC/DSCResources/MSFT_TeamsFederationConfiguration/MSFT_TeamsFederationConfiguration.psm1
@@ -23,11 +23,6 @@ function Get-TargetResource
         [System.Boolean]
         $AllowFederatedUsers,
 
-        # DEPRECATED
-        [Parameter()]
-        [System.Boolean]
-        $AllowPublicUsers,
-
         [Parameter()]
         [System.Boolean]
         $AllowTeamsConsumer,
@@ -130,8 +125,6 @@ function Get-TargetResource
             AllowedDomains                              = $AllowedDomainsValues
             BlockedDomains                              = $BlockedDomainsValues
             AllowFederatedUsers                         = $config.AllowFederatedUsers
-            #DEPRECATED
-            #AllowPublicUsers                            = $config.AllowPublicUsers
             AllowTeamsConsumer                          = $config.AllowTeamsConsumer
             AllowTeamsConsumerInbound                   = $config.AllowTeamsConsumerInbound
             ExternalAccessWithTrialTenants              = $config.ExternalAccessWithTrialTenants
@@ -179,10 +172,6 @@ function Set-TargetResource
         [Parameter()]
         [System.Boolean]
         $AllowFederatedUsers,
-
-        [Parameter()]
-        [System.Boolean]
-        $AllowPublicUsers,
 
         [Parameter()]
         [System.Boolean]
@@ -269,12 +258,6 @@ function Set-TargetResource
         $SetParams.Add('AllowedDomains', $AllowAllKnownDomains)
     }
 
-    if ($SetParams.ContainsKey('AllowPublicUsers'))
-    {
-        Write-Verbose -Message "[DEPRECATED] The AllowPublicUsers property is deprecated and will be removed."
-        $SetParams.Remove('AllowPublicUsers') | Out-Null
-    }
-
     Write-Verbose -Message "SetParams: $(Convert-M365DscHashtableToString -Hashtable $SetParams)"
     Set-CsTenantFederationConfiguration @SetParams
 }
@@ -301,10 +284,6 @@ function Test-TargetResource
         [Parameter()]
         [System.Boolean]
         $AllowFederatedUsers,
-
-        [Parameter()]
-        [System.Boolean]
-        $AllowPublicUsers,
 
         [Parameter()]
         [System.Boolean]
@@ -375,11 +354,6 @@ function Test-TargetResource
     Write-Verbose -Message "Target Values: $(Convert-M365DscHashtableToString -Hashtable $PSBoundParameters)"
 
     $ValuesToCheck = $PSBoundParameters
-    if ($ValuesToCheck.ContainsKey('AllowPublicUsers'))
-    {
-        $ValuesToCheck.Remove('AllowPublicUsers') | Out-Null
-    }
-
     $TestResult = Test-M365DSCParameterState -CurrentValues $CurrentValues `
         -Source $($MyInvocation.MyCommand.Source) `
         -DesiredValues $PSBoundParameters `