+++ Modules/Microsoft365DSC/DSCResources/MSFT_TeamsMeetingPolicy/MSFT_TeamsMeetingPolicy.psm1
@@ -236,10 +236,6 @@ function Get-TargetResource
 
         [Parameter()]
         [System.String]
-        $ForceStreamingAttendeeMode,
-
-        [Parameter()]
-        [System.String]
         $InfoShownInReportMode,
 
         [Parameter()]
@@ -744,10 +740,6 @@ function Set-TargetResource
 
         [Parameter()]
         [System.String]
-        $ForceStreamingAttendeeMode,
-
-        [Parameter()]
-        [System.String]
         $InfoShownInReportMode,
 
         [Parameter()]
@@ -910,9 +902,6 @@ function Set-TargetResource
     $SetParameters.Remove('ManagedIdentity') | Out-Null
     $SetParameters.Remove('Verbose') | Out-Null # Needs to be implicitly removed for the cmdlet to work
     $SetParameters.Remove('AccessTokens') | Out-Null
-
-    # Parameter is Deprecated
-    $SetParameters.Remove('ForceStreamingAttendeeMode') | Out-Null
 
     if ($AllowCloudRecording -eq $false -and $SetParameters.Keys -contains 'AllowRecordingStorageOutsideRegion')
     {
@@ -1196,10 +1185,6 @@ function Test-TargetResource
 
         [Parameter()]
         [System.String]
-        $ForceStreamingAttendeeMode,
-
-        [Parameter()]
-        [System.String]
         $InfoShownInReportMode,
 
         [Parameter()]
@@ -1363,9 +1348,6 @@ function Test-TargetResource
 
     # The AllowUserToJoinExternalMeeting doesn't do anything based on official documentation
     $ValuesToCheck.Remove('AllowUserToJoinExternalMeeting') | Out-Null
-
-    # Parameter is Deprecated
-    $ValuesToCheck.Remove('ForceStreamingAttendeeMode') | Out-Null
 
     if ($AllowCloudRecording -eq $false -and $ValuesToCheck.Keys -contains 'AllowRecordingStorageOutsideRegion')
     {