+++ Modules/Microsoft365DSC/DSCResources/MSFT_TeamsShiftsPolicy/MSFT_TeamsShiftsPolicy.psm1
@@ -24,10 +24,6 @@ function Get-TargetResource
         $EnableScheduleOwnerPermissions,
 
         [Parameter()]
-        [System.Boolean]
-        $EnableShiftPresence,
-
-        [Parameter()]
         [ValidateSet('Always', 'ShowOnceOnChange', 'Never')]
         [System.String]
         $ShiftNoticeFrequency,
@@ -111,8 +107,6 @@ function Get-TargetResource
             AccessGracePeriodMinutes       = $instance.AccessGracePeriodMinutes
             AccessType                     = $instance.AccessType
             EnableScheduleOwnerPermissions = $instance.EnableScheduleOwnerPermissions
-            # DEPRECATED
-            #EnableShiftPresence            = $instance.EnableShiftPresence
             ShiftNoticeFrequency           = $instance.ShiftNoticeFrequency
             ShiftNoticeMessageCustom       = $instance.ShiftNoticeMessageCustom
             ShiftNoticeMessageType         = $instance.ShiftNoticeMessageType
@@ -159,10 +153,6 @@ function Set-TargetResource
         [Parameter()]
         [System.Boolean]
         $EnableScheduleOwnerPermissions,
-
-        [Parameter()]
-        [System.Boolean]
-        $EnableShiftPresence,
 
         [Parameter()]
         [ValidateSet('Always', 'ShowOnceOnChange', 'Never')]
@@ -234,12 +224,6 @@ function Set-TargetResource
     $PSBoundParameters.Remove('ManagedIdentity') | Out-Null
     $PSBoundParameters.Remove('AccessTokens') | Out-Null
 
-    if ($PSBoundParameters.ContainsKey('EnableShiftPresence'))
-    {
-        Write-Verbose -Message 'The EnableShiftPresence parameter was used but is deprecated. It will be ignored.'
-        $PSBoundParameters.Remove('EnableShiftPresence') | Out-Null
-    }
-
     if ($Ensure -eq 'Present' -and $currentInstance.Ensure -eq 'Absent')
     {
         $CreateParameters = ([Hashtable]$PSBoundParameters).Clone()
@@ -310,10 +294,6 @@ function Test-TargetResource
         $EnableScheduleOwnerPermissions,
 
         [Parameter()]
-        [System.Boolean]
-        $EnableShiftPresence,
-
-        [Parameter()]
         [ValidateSet('Always', 'ShowOnceOnChange', 'Never')]
         [System.String]
         $ShiftNoticeFrequency,
@@ -374,9 +354,6 @@ function Test-TargetResource
     $CurrentValues = Get-TargetResource @PSBoundParameters
     $ValuesToCheck = ([Hashtable]$PSBoundParameters).Clone()
     $ValuesToCheck.Remove('Identity') | Out-Null
-    $ValuesToCheck.Remove('EnableShiftPresence') | Out-Null
-
-
 
     Write-Verbose -Message "Current Values: $(Convert-M365DscHashtableToString -Hashtable $CurrentValues)"
     Write-Verbose -Message "Target Values: $(Convert-M365DscHashtableToString -Hashtable $ValuesToCheck)"