+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.AADAuthenticationMethodPolicy.Tests.ps1
@@ -50,7 +50,6 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     Description = "FakeStringValue"
                     DisplayName = "FakeStringValue"
                     Id = "FakeStringValue"
-                    PolicyMigrationState = "preMigration"
                     PolicyVersion = "FakeStringValue"
                     ReconfirmationInDays = 25
                     RegistrationEnforcement = (New-CimInstance -ClassName MSFT_MicrosoftGraphregistrationEnforcement -Property @{
@@ -117,7 +116,6 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     Description = "FakeStringValue"
                     DisplayName = "FakeStringValue"
                     Id = "FakeStringValue"
-                    PolicyMigrationState = "preMigration"
                     PolicyVersion = "FakeStringValue"
                     ReconfirmationInDays = 25
                     RegistrationEnforcement = (New-CimInstance -ClassName MSFT_MicrosoftGraphregistrationEnforcement -Property @{
@@ -236,7 +234,6 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     Description = "FakeStringValue"
                     DisplayName = "FakeStringValue"
                     Id = "FakeStringValue"
-                    PolicyMigrationState = "preMigration"
                     PolicyVersion = "FakeStringValue"
                     ReconfirmationInDays = 25
                     RegistrationEnforcement = (New-CimInstance -ClassName MSFT_MicrosoftGraphregistrationEnforcement -Property @{