+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.EXOHostedContentFilterPolicy.Tests.ps1
@@ -88,12 +88,9 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     HighConfidencePhishAction            = 'Quarantine'
                     HighConfidenceSpamAction             = 'Quarantine'
                     SpamAction                           = 'MoveToJmf'
-                    EnableEndUserSpamNotifications       = $true
                     DownloadLink                         = $false
                     EnableRegionBlockList                = $true
                     EnableLanguageBlockList              = $true
-                    EndUserSpamNotificationCustomSubject = 'This is SPAM'
-                    EndUserSpamNotificationLanguage      = 'Default'
                     BulkThreshold                        = 5
                     AllowedSenders                       = @('test@contoso.com', 'test@fabrikam.com')
                     AllowedSenderDomains                 = @('contoso.com', 'fabrikam.com')
@@ -135,7 +132,6 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     RedirectToRecipients                 = @()
                     TestModeBccToRecipients              = @()
                     QuarantineRetentionPeriod            = 15
-                    EndUserSpamNotificationFrequency     = 1
                     TestModeAction                       = 'AddXHeader'
                     IncreaseScoreWithImageLinks          = 'Off'
                     IncreaseScoreWithNumericIps          = 'On'
@@ -159,12 +155,9 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     HighConfidencePhishAction            = 'Quarantine'
                     HighConfidenceSpamAction             = 'Quarantine'
                     SpamAction                           = 'MoveToJmf'
-                    EnableEndUserSpamNotifications       = $true
                     DownloadLink                         = $false
                     EnableRegionBlockList                = $true
                     EnableLanguageBlockList              = $true
-                    EndUserSpamNotificationCustomSubject = 'This is SPAM'
-                    EndUserSpamNotificationLanguage      = 'Default'
                     BulkThreshold                        = 5
                     AllowedSenders                       = @('test@contoso.com', 'test@fabrikam.com')
                     AllowedSenderDomains                 = @('contoso.com', 'fabrikam.com')
@@ -261,7 +254,6 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     RedirectToRecipients                 = @()
                     TestModeBccToRecipients              = @()
                     QuarantineRetentionPeriod            = 15
-                    EndUserSpamNotificationFrequency     = 1
                     TestModeAction                       = 'AddXHeader'
                     IncreaseScoreWithImageLinks          = 'Off'
                     IncreaseScoreWithNumericIps          = 'On'
@@ -285,12 +277,9 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     HighConfidencePhishAction            = 'Quarantine'
                     HighConfidenceSpamAction             = 'Quarantine'
                     SpamAction                           = 'MoveToJmf'
-                    EnableEndUserSpamNotifications       = $true
                     DownloadLink                         = $false
                     EnableRegionBlockList                = $true
                     EnableLanguageBlockList              = $true
-                    EndUserSpamNotificationCustomSubject = 'This is SPAM'
-                    EndUserSpamNotificationLanguage      = 'Default'
                     BulkThreshold                        = 5
                     AllowedSenders                       = @('test@contoso.com', 'test@fabrikam.com')
                     AllowedSenderDomains                 = @('contoso.com', 'fabrikam.com')